kvformat: rename from welf
As this reflects the purpose of this module much better, WELF is just
one of the format it has support for.
