+++ modules/kvformat/tests/Makefile.am
@@ -1,0 +1,26 @@
+modules_kvformat_tests_TESTS		= \
+	modules/kvformat/tests/test_format_welf	\
+	modules/kvformat/tests/test_kv_scanner	\
+	modules/kvformat/tests/test_kv_parser
+
+check_PROGRAMS				+= ${modules_kvformat_tests_TESTS}
+
+modules_kvformat_tests_test_format_welf_CFLAGS	= $(TEST_CFLAGS)
+modules_kvformat_tests_test_format_welf_LDADD	= $(TEST_LDADD)
+modules_kvformat_tests_test_format_welf_LDFLAGS	= \
+	$(PREOPEN_SYSLOGFORMAT)		  \
+	-dlpreopen $(top_builddir)/modules/kvformat/libkvformat.la
+modules_kvformat_tests_test_format_welf_DEPENDENCIES = $(top_builddir)/modules/kvformat/libkvformat.la
+
+modules_kvformat_tests_test_kv_parser_CFLAGS	= $(TEST_CFLAGS) -I$(top_srcdir)/modules/kvformat
+modules_kvformat_tests_test_kv_parser_LDADD	= $(TEST_LDADD)
+modules_kvformat_tests_test_kv_parser_LDFLAGS	= \
+	$(PREOPEN_SYSLOGFORMAT)		  \
+	-dlpreopen $(top_builddir)/modules/kvformat/libkvformat.la
+modules_kvformat_tests_test_kv_parser_DEPENDENCIES = $(top_builddir)/modules/kvformat/libkvformat.la
+
+modules_kvformat_tests_test_kv_scanner_CFLAGS	= $(TEST_CFLAGS) -I$(top_srcdir)/modules/kvformat
+modules_kvformat_tests_test_kv_scanner_LDADD	= $(TEST_LDADD)
+modules_kvformat_tests_test_kv_scanner_LDFLAGS	= \
+	-dlpreopen $(top_builddir)/modules/kvformat/libkvformat.la
+modules_kvformat_tests_test_kv_scanner_DEPENDENCIES = $(top_builddir)/modules/kvformat/libkvformat.la
