+++ modules/kvformat/tests/test_format_welf.c
@@ -1,5 +1,6 @@
 /*
- * Copyright (c) 2015 BalaBit
+ * Copyright (c) 2011-2013 BalaBit IT Ltd, Budapest, Hungary
+ * Copyright (c) 2011-2013 Gergely Nagy <algernon@balabit.hu>
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License version 2 as published
@@ -19,36 +20,30 @@
  * COPYING for details.
  *
  */
-
+#include "template_lib.h"
+#include "apphook.h"
 #include "plugin.h"
-#include "plugin-types.h"
-#include "format-welf.h"
+#include "cfg.h"
 
-extern CfgParser kv_parser_parser;
-
-static Plugin welf_plugins[] =
+void
+test_format_welf(void)
 {
-  {
-    .type = LL_CONTEXT_PARSER,
-    .name = "kv-parser",
-    .parser = &kv_parser_parser,
-  },
-  TEMPLATE_FUNCTION_PLUGIN(tf_format_welf, "format-welf"),
-};
-
-gboolean
-welf_module_init(GlobalConfig *cfg, CfgArgs *args)
-{
-  plugin_register(cfg, welf_plugins, G_N_ELEMENTS(welf_plugins));
-  return TRUE;
+  assert_template_format("$(format-welf MSG=$MSG)", "MSG=árvíztűrőtükörfúrógép");
+  assert_template_format("$(format-welf MSG=$escaping)", "MSG=\"binary stuff follows \\\"\\xad árvíztűrőtükörfúrógép\"");
+  assert_template_format_with_context("$(format-welf MSG=$MSG)", "MSG=árvíztűrőtükörfúrógép MSG=árvíztűrőtükörfúrógép");
 }
 
-const ModuleInfo module_info =
+int
+main(int argc G_GNUC_UNUSED, char *argv[] G_GNUC_UNUSED)
 {
-  .canonical_name = "welf",
-  .version = VERSION,
-  .description = "The welf module provides WebTrends Enhanced Log Format support for syslog-ng.",
-  .core_revision = SOURCE_REVISION,
-  .plugins = welf_plugins,
-  .plugins_len = G_N_ELEMENTS(welf_plugins),
-};
+  app_startup();
+  putenv("TZ=UTC");
+  tzset();
+  init_template_tests();
+  plugin_load_module("kvformat", configuration, NULL);
+
+  test_format_welf();
+
+  deinit_template_tests();
+  app_shutdown();
+}
